refactor(providers): rename ValidateResponse to HandleResponseError
Make handleResponseError abstract
