+++ src/Providers/DeepSeek/Handlers/Structured.php
@@ -4,9 +4,8 @@ namespace Prism\Prism\Providers\DeepSeek\Handlers;
 
 use Illuminate\Http\Client\PendingRequest;
 use Illuminate\Support\Arr;
-use Prism\Prism\Exceptions\PrismException;
+use Prism\Prism\Providers\DeepSeek\Concerns\HandleResponseError;
 use Prism\Prism\Providers\DeepSeek\Concerns\MapsFinishReason;
-use Prism\Prism\Providers\DeepSeek\Concerns\ValidatesResponses;
 use Prism\Prism\Providers\DeepSeek\Maps\FinishReasonMap;
 use Prism\Prism\Providers\DeepSeek\Maps\MessageMap;
 use Prism\Prism\Structured\Request;
@@ -20,8 +19,8 @@ use Prism\Prism\ValueObjects\Usage;
 
 class Structured
 {
+    use HandleResponseError;
     use MapsFinishReason;
-    use ValidatesResponses;
 
     protected ResponseBuilder $responseBuilder;
 
@@ -36,7 +35,7 @@ class Structured
 
         $data = $this->sendRequest($request);
 
-        $this->validateResponse($data);
+        $this->handleResponseError();
 
         return $this->createResponse($request, $data);
     }
@@ -46,7 +45,7 @@ class Structured
      */
     protected function sendRequest(Request $request): array
     {
-        $response = $this->client->post(
+        $this->httpResponse = $this->client->post(
             'chat/completions',
             array_merge([
                 'model' => $request->model(),
@@ -59,17 +58,7 @@ class Structured
             ]))
         );
 
-        return $response->json();
-    }
-
-    /**
-     * @param  array<string, mixed>  $data
-     */
-    protected function validateResponse(array $data): void
-    {
-        if ($data === []) {
-            throw PrismException::providerResponseError('DeepSeek Error: Empty response');
-        }
+        return $this->httpResponse->json();
     }
 
     /**
