+++ src/Providers/Gemini/Concerns/HandleResponseError.php
@@ -2,31 +2,27 @@
 
 declare(strict_types=1);
 
-namespace Prism\Prism\Providers\Groq\Concerns;
+namespace Prism\Prism\Providers\Gemini\Concerns;
 
 use Illuminate\Http\Client\Response;
 use Prism\Prism\Exceptions\PrismException;
-use Prism\Prism\ValueObjects\ProviderRateLimit;
 
-trait ValidateResponse
+trait HandleResponseError
 {
-    protected function validateResponse(): void
+    protected Response $httpResponse;
+
+    protected function handleResponseError(): void
     {
         $data = $this->httpResponse->json();
 
         if (! $data || data_get($data, 'error')) {
             throw PrismException::providerResponseError(vsprintf(
-                'Groq Error:  [%s] %s',
+                'Gemini Error: [%s] %s',
                 [
-                    data_get($data, 'error.type', 'unknown'),
+                    data_get($data, 'error.code', 'unknown'),
                     data_get($data, 'error.message', 'unknown'),
                 ]
             ));
         }
     }
-
-    /**
-     * @return ProviderRateLimit[]
-     */
-    abstract protected function processRateLimits(Response $response): array;
 }
