+++ src/Providers/Groq/Concerns/HandleResponseError.php
@@ -2,19 +2,23 @@
 
 declare(strict_types=1);
 
-namespace Prism\Prism\Providers\XAI\Concerns;
+namespace Prism\Prism\Providers\Groq\Concerns;
 
+use Illuminate\Http\Client\Response;
 use Prism\Prism\Exceptions\PrismException;
+use Prism\Prism\ValueObjects\ProviderRateLimit;
 
-trait ValidatesResponses
+trait HandleResponseError
 {
-    protected function validateResponse(): void
+    protected Response $httpResponse;
+
+    protected function handleResponseError(): void
     {
         $data = $this->httpResponse->json();
 
         if (! $data || data_get($data, 'error')) {
             throw PrismException::providerResponseError(vsprintf(
-                'XAI Error:  [%s] %s',
+                'Groq Error:  [%s] %s',
                 [
                     data_get($data, 'error.type', 'unknown'),
                     data_get($data, 'error.message', 'unknown'),
@@ -22,4 +26,9 @@ trait ValidatesResponses
             ));
         }
     }
+
+    /**
+     * @return ProviderRateLimit[]
+     */
+    abstract protected function processRateLimits(Response $response): array;
 }
