+++ src/Providers/Groq/Handlers/Structured.php
@@ -5,8 +5,8 @@ namespace Prism\Prism\Providers\Groq\Handlers;
 use Illuminate\Http\Client\PendingRequest;
 use Illuminate\Http\Client\Response as ClientResponse;
 use Illuminate\Support\Arr;
+use Prism\Prism\Providers\Groq\Concerns\HandleResponseError;
 use Prism\Prism\Providers\Groq\Concerns\ProcessRateLimits;
-use Prism\Prism\Providers\Groq\Concerns\ValidateResponse;
 use Prism\Prism\Providers\Groq\Maps\FinishReasonMap;
 use Prism\Prism\Providers\Groq\Maps\MessageMap;
 use Prism\Prism\Structured\Request;
@@ -20,7 +20,7 @@ use Prism\Prism\ValueObjects\Usage;
 
 class Structured
 {
-    use ProcessRateLimits, ValidateResponse;
+    use HandleResponseError, ProcessRateLimits;
 
     protected ClientResponse $httpResponse;
 
@@ -37,7 +37,7 @@ class Structured
 
         $response = $this->sendRequest($request);
 
-        $this->validateResponse();
+        $this->handleResponseError();
 
         $data = $response->json();
